requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.89.1
flake8==6.0.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
orjson==3.8.5
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
